linux-gen: pktio: move common code out of socket pktio files
Cleanup socket pktio code by moving common functionality to a
separate file.
