+++ platform/linux-generic/Makefile.am
@@ -112,6 +112,7 @@ noinst_HEADERS = \
 		  include/odp_packet_io_ipc_internal.h \
 		  include/odp_packet_io_ring_internal.h \
 		  include/odp_packet_socket.h \
+		  include/odp_socket_common.h \
 		  include/odp_ethtool_stats.h \
 		  include/odp_ethtool_rss.h \
 		  include/odp_pcapng.h \
@@ -201,6 +202,7 @@ __LIB__libodp_linux_la_SOURCES = \
 			   odp_weak.c \
 			   pktio/dpdk.c \
 			   pktio/dpdk_parse.c \
+			   pktio/socket_common.c \
 			   pktio/ethtool_stats.c \
 			   pktio/ethtool_rss.c \
 			   pktio/io_ops.c \
