+++ platform/linux-generic/pktio/socket_mmap.c
@@ -28,6 +28,7 @@
 #include <odp_api.h>
 #include <odp/api/plat/packet_inlines.h>
 #include <odp_packet_socket.h>
+#include <odp_socket_common.h>
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
 #include <odp_ethtool_stats.h>
