+++ platform/linux-generic/pktio/tap.c
@@ -44,6 +44,7 @@
 #include <odp_api.h>
 #include <odp/api/plat/packet_inlines.h>
 #include <odp_packet_socket.h>
+#include <odp_socket_common.h>
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
 #include <odp_classification_internal.h>
